refactor : 페어매치 출력분 클래스 분리
