+++ src/main/java/userInterafce/PairMatch.java
@@ -10,7 +10,6 @@ import pairmatching.Course;
 import pairmatching.Crew;
 import pairmatching.Crews;
 import pairmatching.Level;
-import pairmatching.Mission;
 import pairmatching.PairStorage;
 
 public class PairMatch {
@@ -20,10 +19,11 @@ public class PairMatch {
 	}
 
 	public void run() {
-		new printCourseAndMission().print();
+		new PrintInterface().printCourseAndMission();
 	}
 
 	public void matchOneByOne(String name) {
+
 
 	}
 
@@ -53,39 +53,4 @@ public class PairMatch {
 		return crewNameList;
 	}
 
-	class printCourseAndMission {
-		public void print() {
-			String lineDistinguish = "#############################################";
-			System.out.println(lineDistinguish);
-			printCourse();
-			printMission();
-			System.out.println(lineDistinguish);
-		}
-
-		private void printCourse() {
-			List<String> courseNameList = new ArrayList<>();
-			for (Course course : Course.values()) {
-				courseNameList.add(course.toString());
-			}
-			String courseListAsString = String.join(" | ", courseNameList);
-			System.out.printf("과정: %s\n", courseListAsString);
-		}
-
-		private void printMission() {
-			System.out.println("미션:");
-			for (Level level:Level.values()) {
-				printMissionByLevel(level);
-			}
-		}
-
-		private void printMissionByLevel(Level level) {
-			List<String> missionNameList = new ArrayList<>();
-			for (Mission mission : Mission.getMissionListByLevel(level)) {
-				missionNameList.add(mission.toString());
-			}
-
-			String courseListAsString = String.join(" | ", missionNameList);
-			System.out.printf("  - %s: %s\n",level.toString(), courseListAsString);
-		}
-	}
 }
